change pulsarrecord and fix tracker datatype in pulsario
